external/optee_nuttx: add compat implementation for depolyed devices
for these deployed devices, the huk and chip id fetch procedure is bound
together, for example:
if the huk id that fetched is 0xff, then the chip id need to set as
"0xbeef",
if the huk id is not filled with 0xff, then the chip id need also using
the true value.
This mechanism is different from latest implementation, of which the huk
is not bound with chip id.

So In order to compat with the deployed device(specific product/specific product), we need to add
an extra param to the tee_otp_get_hw_unique_key and tee_otp_get_die_id_compat
method to implement compatibility.

Signed-off-by: guoshichao <guoshichao@xiaomi.com>