optee_nuttx: make the huk length compat with deployed device
In devices that have already been deployed by using mitee, the
length of HUK is fixed at 16. If we set the length of HUK to be
the same as CONFIG_BOARDCTL_UNIQUEKEY_SIZE, then the FEK constructed
based on this HUK will not be able to decrypt the pre-set "TRIAD"
in the devices that have already been implemented using mitee.
Therefore, for devices that have already been released, it is
recommended to directly set the length of HWUNIQUE_KEY_LENGTH to 16
to ensure that they can work properly and avoid the problem of
being unable to decrypt the "TRIAD" in already deployed devices
that using mitee.

Signed-off-by: guoshichao <guoshichao@xiaomi.com>