optee_nuttx: fix the busy-loop bug of condvar_wait in tee
In TEE, all the underlying implementations involving condvar
and mutex are based on some implementations in wait_queue.c.
There is a crucial function in it called wq_wait_final().
The purpose of this function is mainly to wait for the lock
to be released.

When we adjust the log trace_level in Tee from TRACE_DEBUG to
TRACE_ERROR, the implementation of wp_wait_final() basically becomes
an empty function.
In the case of TRACE_DEBUG log level, the wp_wait_final() function
will print logs, and the current Tee log implementation is based
on syslog-rpmsg, which means printing logs will perform a
switch to REE, and this yield current cpu from TEE to REE.
If the log level is changed to TRACE_ERROR, wp_wait_final() will be
basically an empty implementation, and wp_wait_final() will keep
waiting for the lock to be released(make current thread busy-loop),
and blocking the current execution flow, thus causing TEE's execution
to stall.
As the core part of wp_wait_final() is notif_wait() function, so we
could re-implement notif_wait() function, and yield current cpu out
in notif_wait() function.

neededby: 4055816
Signed-off-by: guoshichao <guoshichao@xiaomi.com>