user_ta_wasm: fix the Re-opened session misuse wasm function instance
When a TA is requested to be reopened while the previous session has not ended,
the ref_count of this TA will be incremented directly in
tee_ta_manager.c#tee_ta_init_session_with_context(). At this time, the ref_count of
this TA will become 2.
When a function in WASM is called for this new TA session, it is possible that the
WASM function already found in the previous session will be called.
For example, the utc_func maintained by old_TA_session is wasm_B at this time.
Then, the utc_func that new_TA_session hopes to call is wasm_A. Due to the judgment of
ref_count here, new_TA_session will directly reuse the wasm_B function of old_TA_session.
As a result, a problem of incorrect parameters occurs.

Signed-off-by: pengyinjie <pengyinjie@xiaomi.com>