optee_vela: fix the object_id management bug on 64bit mode
In the old implementation, there was a direct mapping between
`object_id` and `object_ptr`. Since `object_id` is 32-bit,
this worked correctly on 32-bit systems.
However, when running on a 64-bit system, the `object_id`
gets truncated to 32 bits, causing data loss and preventing
proper matching between `object_id` and `object_ptr`.
As a result, storage APIs like `TEE_CreatePersistentObject`
malfunction.
To fix this issue on 64-bit systems, we need to refer to
OP-TEE OS's original implementation and implement the
correct mapping instead of using a direct mapping.

Signed-off-by: guoshichao <guoshichao@xiaomi.com>